api/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
PyYAML>=6.0.0
